CSS File
This piece of CSS code outlines the styles for a website's design. It sets the font family, background color, and other settings for the body element. A background image and a logo are used to decorate the header section. A list of links with hover effects can be found in the navigation bar, which is permanently positioned at the top of the page. Two columns, one for an image and the other for a description, make up the "about" section. Tabs that flip between the various contents are also included in the section. Flexbox is used in the code to produce a responsive layout that adapts to various screen sizes.